tests: move stories to dedicated folder and to storybook v 7 format
