+++ typescript/packages/subsurface-viewer/src/layers/axes/boxLayer.ts
