+++ typescript/packages/subsurface-viewer/src/layers/colormap/colormapLayer.ts
