+++ typescript/packages/subsurface-viewer/src/layers/grid3d/test_data/TruncatedSnubCube.ts
