+++ typescript/packages/subsurface-viewer/src/layers/hillshading2d/hillshading2dLayer.ts
