+++ typescript/packages/subsurface-viewer/src/layers/northarrow/northArrow3DLayer.ts
