+++ typescript/packages/subsurface-viewer/src/layers/shader_modules/decoder.ts
