+++ typescript/packages/subsurface-viewer/src/storybook/examples/CameraControlExamples.stories.tsx
@@ -14,13 +14,15 @@ import type { ViewStateType, BoundingBox3D } from "../../components/Map";
 import { AxesLayer } from "../../layers";
 
 import {
-    mainStyle as defaultmainStyle,
+    mainStyle,
     huginAxes3DLayer,
+    default2DViews,
     default3DViews,
+    defaultStoryParameters,
     hugin3DBounds,
-    huginMapDepthPropLayerPng,
-    huginMapNetmapPropLayerPng,
-    huginMeshMapLayerPng,
+    hugin25mDepthMapLayer,
+    hugin25mKhNetmapMapLayer,
+    hugin25mKhNetmapMapLayerPng,
     northArrowLayer,
     volveWellsBounds,
     volveWellsLayer,
@@ -38,20 +40,13 @@ const stories: Meta = {
 };
 export default stories;
 
-const defaultStoryParameters = {
-    docs: {
-        inlineStories: false,
-        iframeHeight: 500,
-    },
-};
-
 const classes = {
     main: "default-main",
     mainWithButton: "mian-with-button",
 };
 
 const Root = styled("div")({
-    ...defaultmainStyle,
+    ...mainStyle,
     [`& .${classes.mainWithButton}`]: {
         height: 500,
         border: "1px solid black",
@@ -115,8 +110,8 @@ const SyncedMultiViewComponent = (args: {
     const subsurfaceViewerArgs: SubsurfaceViewerProps = {
         id: "view_state_synchronization",
         layers: [
-            huginMapNetmapPropLayerPng,
-            huginMapDepthPropLayerPng,
+            hugin25mKhNetmapMapLayer,
+            hugin25mDepthMapLayer,
             volveWellsLayer,
         ],
         views: {
@@ -124,13 +119,13 @@ const SyncedMultiViewComponent = (args: {
             viewports: [
                 {
                     id: "view_1",
-                    layerIds: [huginMapDepthPropLayerPng.id],
+                    layerIds: [hugin25mDepthMapLayer.id],
                     show3D: args.show3d,
                     isSync: args.sync.includes("view_1"),
                 },
                 {
                     id: "view_2",
-                    layerIds: [huginMapNetmapPropLayerPng.id],
+                    layerIds: [hugin25mKhNetmapMapLayer.id],
                     show3D: args.show3d,
                     isSync: args.sync.includes("view_2"),
                 },
@@ -142,10 +137,7 @@ const SyncedMultiViewComponent = (args: {
                 },
                 {
                     id: "view_4",
-                    layerIds: [
-                        volveWellsLayer.id,
-                        huginMapDepthPropLayerPng.id,
-                    ],
+                    layerIds: [volveWellsLayer.id, hugin25mDepthMapLayer.id],
                     show3D: args.show3d,
                     isSync: args.sync.includes("view_4"),
                 },
@@ -235,10 +227,7 @@ export const SyncedSubsurfaceViewers: StoryObj<
         bounds: volveWellsBounds,
         layers: [volveWellsLayer],
         cameraPosition,
-        views: {
-            layout: [1, 1],
-            viewports: [{ id: "view", show3D: false }],
-        },
+        views: default2DViews,
     },
     render: (args) => <SyncedCameraSettingsComponent {...args} />,
 };
@@ -321,8 +310,7 @@ export const AutoZoomToBoxStory: StoryObj<typeof SubsurfaceViewer> = {
     },
     parameters: {
         docs: {
-            inlineStories: false,
-            iframeHeight: 500,
+            ...defaultStoryParameters.docs,
             description: {
                 story: "",
             },
@@ -334,14 +322,14 @@ export const AutoZoomToBoxStory: StoryObj<typeof SubsurfaceViewer> = {
 const ScaleZComponent: React.FC<SubsurfaceViewerProps> = (args) => {
     const [layers, setLayers] = React.useState([
         huginAxes3DLayer,
-        huginMeshMapLayerPng,
+        hugin25mKhNetmapMapLayerPng,
         northArrowLayer,
     ]);
 
     const handleChange = () => {
         setLayers([
             huginAxes3DLayer,
-            huginMeshMapLayerPng,
+            hugin25mKhNetmapMapLayerPng,
             volveWellsWithLogsLayer,
             northArrowLayer,
         ]);
@@ -367,7 +355,7 @@ export const ScaleZ: StoryObj<typeof SubsurfaceViewer> = {
         id: "ScaleZ",
         layers: [
             huginAxes3DLayer,
-            huginMeshMapLayerPng,
+            hugin25mKhNetmapMapLayerPng,
             volveWellsWithLogsLayer,
             northArrowLayer,
         ],
@@ -380,7 +368,7 @@ export const ScaleZ: StoryObj<typeof SubsurfaceViewer> = {
                     id: "view_1",
                     layerIds: [
                         huginAxes3DLayer.id,
-                        huginMeshMapLayerPng.id,
+                        hugin25mKhNetmapMapLayerPng.id,
                         northArrowLayer.id,
                     ],
                     show3D: true,
@@ -447,7 +435,11 @@ const ResetCameraComponent: React.FC<SubsurfaceViewerProps> = (args) => {
 export const ResetCameraStory: StoryObj<typeof SubsurfaceViewer> = {
     args: {
         id: "ResetCameraProperty",
-        layers: [huginAxes3DLayer, huginMeshMapLayerPng, northArrowLayer],
+        layers: [
+            huginAxes3DLayer,
+            hugin25mKhNetmapMapLayerPng,
+            northArrowLayer,
+        ],
 
         bounds: volveWellsBounds,
         cameraPosition: ResetCameraPropertyDefaultCameraPosition,
@@ -467,13 +459,13 @@ export const ResetCameraStory: StoryObj<typeof SubsurfaceViewer> = {
 const AddLayerComponent: React.FC<SubsurfaceViewerProps> = (args) => {
     const [layers, setLayers] = React.useState([
         huginAxes3DLayer,
-        huginMeshMapLayerPng,
+        hugin25mKhNetmapMapLayerPng,
         northArrowLayer,
     ]);
 
     const handleChange = () => {
         setLayers([
-            huginMeshMapLayerPng,
+            hugin25mKhNetmapMapLayerPng,
             huginAxes3DLayer,
             volveWellsWithLogsLayer,
             northArrowLayer,
